fix: update font asset and fix UI scope handling and first 7.5 compatiblity update - Change JP-specific font to CJK font for broader character support - Fix variable scope in DataAnalysisUi to prevent premature disposal - Update zone tracking variable type from ushort to uint for compatibility - Bump DalamudPackager dependency to latest version

+++ Sphene/Services/DalamudUtilService.cs
@@ -48,7 +48,7 @@ public class DalamudUtilService : IHostedService, IMediatorSubscriber
     private DateTime _delayedFrameworkUpdateCheck = DateTime.UtcNow;
     private string _lastGlobalBlockPlayer = string.Empty;
     private string _lastGlobalBlockReason = string.Empty;
-    private ushort _lastZone = 0;
+    private uint _lastZone = 0;
     private readonly Dictionary<string, (string Name, nint Address)> _playerCharas = new(StringComparer.Ordinal);
     private readonly List<string> _notUpdatedCharas = [];
     private bool _sentBetweenAreas = false;
